logs completed

+++ src/graphql/resolvers/notice.resolver.ts
@@ -1,30 +1,47 @@
 import {ApolloError} from "apollo-server";
 import { Notice } from "../../models/notice.model";
-import { Campus } from "../../models/campus.model";
 import {getProjection, transformNotice} from "./merge";
 import {isAuth} from "../../middleware/is-auth";
 import {config} from "../../../enviroments.dev";
+import {
+  registerBadLog,
+  registerGoodLog,
+  registerErrorLog
+} from "../../middleware/logAction";
 
 const noticeQueries = {
   notice: async(_, args, context, info) => {
-    // if (!await isAuth(context, [config.permission.docente]))
-    //   throw new ApolloError('Unauthenticated');
+    const qType = 'Query';
+    const qName = 'notice';
     try {
+      if (!await isAuth(context, [config.permission.docente])) {
+        registerBadLog(context, qType, qName);
+        throw new ApolloError('Error: S5');
+      }
+
       const projections = getProjection(info);
       let doc = await Notice.findById(args.id, projections).exec();
       if (projections.createdBy) {
         // query.populate('createdBy');
         doc = transformNotice(doc);
       }
+      registerGoodLog(context, qType, qName, args.id);
       return doc;
     } catch (e) {
+      registerErrorLog(context, qType, qName);
       throw new ApolloError(e);
     }
   },
   notices: async(_, {page, perPage}, context, info) => {
-    // if (!await isAuth(context, [config.permission.docente]))
-    //   throw new ApolloError('Unauthenticated');
+    const qType = 'Query';
+    const qName = 'notices';
+
     try {
+      if (!await isAuth(context, [config.permission.docente])) {
+        registerBadLog(context, qType, qName);
+        throw new ApolloError('Error: S5');
+      }
+
       const projections = getProjection(info);
       let docs = await Notice
         .find({}, projections)
@@ -32,11 +49,12 @@ const noticeQueries = {
         .limit(perPage).exec();
 
       if (projections.createdBy) {
-        // query.populate('createdBy');
         docs = docs.map(transformNotice);
       }
+      registerGoodLog(context, qType, qName, 'Multiple documents');
       return docs;
     } catch (e) {
+      registerErrorLog(context, qType, qName);
       throw new ApolloError(e);
     }
   }
@@ -44,9 +62,14 @@ const noticeQueries = {
 
 const noticeMutations = {
   createNotice: async(_: any, { input }, context: any, info: any) => {
-    // if (!await isAuth(context, [config.permission.admin]))
-    //   throw new ApolloError('Unauthenticated');
+    const qType = 'Mutation';
+    const qName = 'createNotice';
     try {
+      if (!await isAuth(context, [config.permission.admin])) {
+        registerBadLog(context, qType, qName);
+        throw new ApolloError('Error: S5');
+      }
+
       const notice = new Notice({
         title: input.title,
         body: input.body,
@@ -57,15 +80,23 @@ const noticeMutations = {
         toDate: input.toDate,
         createdBy: input.createdBy
       });
-      return await notice.save();
+      const doc = await notice.save();
+      registerGoodLog(context, qType, qName, doc._id)
+      return doc;
     }catch (e) {
+      registerErrorLog(context, qType, qName);
       throw new ApolloError(e);
     }
   },
   updateNotice: async(_, args, context, info) => {
-    // if (!await isAuth(context, [config.permission.admin]))
-    //   throw new ApolloError('Unauthenticated');
+    const qType = 'Mutation';
+    const qName = 'updateNotice';
     try {
+      if (!await isAuth(context, [config.permission.admin])) {
+        registerBadLog(context, qType, qName);
+        throw new ApolloError('Error: S5');
+      }
+
       const projections = getProjection(info);
       let doc = await Notice
         .findById(args.id, projections)
@@ -74,19 +105,27 @@ const noticeMutations = {
         // query.populate('createdBy');
         doc = transformNotice(doc);
       }
+      registerGoodLog(context, qType, qName, doc._id);
       return doc;
-      // return await Notice.findByIdAndUpdate(args.id, args.input, {new: true}).exec();
     }catch (e) {
+      registerErrorLog(context, qType, qName);
       throw new ApolloError(e);
     }
   },
   deleteNotice: async(_, args, context) => {
-    // if (!await isAuth(context, [config.permission.admin]))
-    //   throw new ApolloError('Unauthenticated');
+    const qType = 'Mutation';
+    const qName = 'deleteNotice';
     try {
-      const res = await Notice.findByIdAndDelete(args.id);
-      return res;
+      if (!await isAuth(context, [config.permission.admin])) {
+        registerBadLog(context, qType, qName);
+        throw new ApolloError('Error: S5');
+      }
+
+      const doc = await Notice.findByIdAndDelete(args.id);
+      registerGoodLog(context, qType, qName, doc._id)
+      return doc;
     } catch (e) {
+      registerErrorLog(context, qType, qName);
       throw new ApolloError(e);
     }
   }
